requirements.txt
Django==4.0.2
django_mptt==0.13.4
djangorestframework==3.13.1
djoser==2.1.0
python-dotenv==0.20.0
psycopg2-binary==2.9.3
django-rest-swagger==2.2.0